+++ voxel_octree/SuperContainer/SparseSuperContainer.hpp
@@ -78,7 +78,7 @@ struct SparseSuperContainer
     /*!
         \brief Returns number of voxels
     */
-    uint16_t            getNbVoxel() const;
+    uint32_t            getNbVoxel() const;
     /*!
         \brief Find voxel
         \param it Iterator containing voxel position information
